added DonAltCapone
My second gamejam project i did in a team of 5. I worked on the enemy AI. I added behaviours so the AI would detect the player in a cone shaped line of sight and would try to chase the player and shoot him, when in range. Sadly the level designer didnt finish the combat levels in time, so the AI wasn't present in the shipped game.